refactor(file): Move FileContent and FilePacket to its own file.
